+++ libsrc/hyperion/PriorityMuxer.cpp
@@ -85,7 +85,7 @@ bool PriorityMuxer::setSourceAutoSelectEnabled(bool enable, bool update)
 	return false;
 }
 
-bool PriorityMuxer::setPriority(uint8_t priority)
+bool PriorityMuxer::setPriority(int priority)
 {
 	if(_activeInputs.contains(priority))
 	{
@@ -216,7 +216,7 @@ bool PriorityMuxer::setInputImage(int priority, const Image<ColorRgb>& image, in
 		return false;
 	}
 
-	// calc final timeout
+	// calculate final timeout
 	if(timeout_ms > 0)
 		timeout_ms = QDateTime::currentMSecsSinceEpoch() + timeout_ms;
 
@@ -248,13 +248,13 @@ bool PriorityMuxer::setInputImage(int priority, const Image<ColorRgb>& image, in
 	return true;
 }
 
-bool PriorityMuxer::setInputInactive(quint8 priority)
+bool PriorityMuxer::setInputInactive(int priority)
 {
 	Image<ColorRgb> image;
 	return setInputImage(priority, image, -100);
 }
 
-bool PriorityMuxer::clearInput(uint8_t priority)
+bool PriorityMuxer::clearInput(int priority)
 {
 	if (priority < PriorityMuxer::LOWEST_PRIORITY && _activeInputs.remove(priority))
 	{
@@ -299,7 +299,7 @@ void PriorityMuxer::setCurrentTime()
 	{
 		if (infoIt->timeoutTime_ms > 0 && infoIt->timeoutTime_ms <= now)
 		{
-			quint8 tPrio = infoIt->priority;
+			int tPrio = infoIt->priority;
 			infoIt = _activeInputs.erase(infoIt);
 			Debug(_log,"Timeout clear for priority %d",tPrio);
 			emit prioritiesChanged();
@@ -317,7 +317,7 @@ void PriorityMuxer::setCurrentTime()
 			++infoIt;
 		}
 	}
-	// eval if manual selected prio is still available
+	// evaluate, if manual selected priority is still available
 	if(!_sourceAutoSelectEnabled)
 	{
 		if(_activeInputs.contains(_manualSelectedPriority))
